switch to triton ascend 3.2.1 (#281)
* modify compiler_hint path to tl.extra.cann.extension; tl.max add propagate_nan=tl.PropagateNan.ALL;

* swa add enable_ubuf_saving to solve ub overflow

* normalization use casting_mode gemma to avoid F.rms_norm fp32 cast precision

* remove print

* flash attention use need_mask to do if else

* Revert "flash attention use need_mask to do if else"

This reverts commit d70513120d56ea5a19ac29f1d460d4d8f331b2e0.

* modify test_over_encoding to switch to triton-ascend

* modify n_gram mask to solve random precision problem

* switch to triton-ascend 3.2.1, use wget temprorarily for checking CI is right

* switch to triton-ascend 3.2.1, use wget temprorarily for checking CI is right

* switch to triton-ascend 3.2.1, use wget temprorarily for checking CI is right

* add print to debug CI

* modify ccec path

* add CI debug

* add CI debug

* add CI debug

* rollback ci and switch to cann8.5.0 image

* add triton-ascend on CI

* fix inder quant para error

* fix perf test case

* modify extract_slice _compute_vision_rope to adapter triton-ascend

* add sync_solver=False to avoid groupgemm perf descend on triton-ascend

* switch to byted-triton-x 3.2.1

* add --index-url to switch to byted-triton-x 3.2.1

* add --index-url to switch to byted-triton-x 3.2.1

* Update mojo_opset/backends/ttx/kernels/npu/over_encoding/n_gram.py


* rollback to rmsnorm_fwd llama mode, triton-ascend 3.2.1 precision is OK

* functions rms backward precision is not OK, change to gemma

* no need to add propagate_nan to all

* groupgemm: delete dot_pad_only_k hint and tl.multibuffer

* modify ci and pyproject

* use pyproject.toml to pip install

---------

+++ mojo_opset/backends/ttx/kernels/npu/rmsnorm.py
@@ -343,7 +343,7 @@ def _rmsnorm_bwd_kernel(
 
         if casting_mode_int == _CASTING_MODE_LLAMA:
             m_block = (dY_block * W_row_offset[None, :]).to(tl.float32)
-            dW_acc += tl.sum(dY_block * normed_X_block.to(X_dtype), axis=0)
+            dW_acc += tl.sum(dY_block * normed_X_block.to(tl.float32), axis=0)
         elif casting_mode_int == _CASTING_MODE_GEMMA:
             dY_block_f32 = dY_block.to(tl.float32)
             W_row_offset = W_row_offset.to(tl.float32)
@@ -442,7 +442,7 @@ def _rmsnorm_bwd_large_cols_kernel(
 
             if casting_mode_int == _CASTING_MODE_LLAMA:
                 m_chunk = (dY_chunk * W_chunk_offset[None, :]).to(tl.float32)
-                dW_chunk_sum = tl.sum(dY_chunk * normed_X_chunk.to(X_dtype), axis=0)
+                dW_chunk_sum = tl.sum(dY_chunk * normed_X_chunk.to(tl.float32), axis=0)
             elif casting_mode_int == _CASTING_MODE_GEMMA:
                 dY_chunk_f32 = dY_chunk.to(tl.float32)
                 W_chunk_offset = W_chunk_offset.to(tl.float32)
